Made hzerrors public
