+++ internal/cluster/connection_bound_invocation.go
@@ -20,11 +20,10 @@ import (
 	"errors"
 	"time"
 
-	"github.com/hazelcast/hazelcast-go-client/internal/cb"
-
-	"github.com/hazelcast/hazelcast-go-client/internal/hzerror"
+	"github.com/hazelcast/hazelcast-go-client/hzerrors"
 
 	pubcluster "github.com/hazelcast/hazelcast-go-client/cluster"
+	"github.com/hazelcast/hazelcast-go-client/internal/cb"
 	"github.com/hazelcast/hazelcast-go-client/internal/invocation"
 	"github.com/hazelcast/hazelcast-go-client/internal/proto"
 )
@@ -55,7 +54,7 @@ func (i *ConnectionBoundInvocation) CanRetry(err error) bool {
 	if errors.Is(err, nonRetryableError) {
 		return false
 	}
-	var ioError *hzerror.HazelcastIOError
-	var targetDisconnectedError *hzerror.HazelcastTargetDisconnectedError
+	var ioError *hzerrors.HazelcastIOError
+	var targetDisconnectedError *hzerrors.HazelcastTargetDisconnectedError
 	return errors.Is(err, ioError) || errors.Is(err, targetDisconnectedError)
 }
