+++ internal/serialization/object_data_test.go
@@ -22,7 +22,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	"github.com/hazelcast/hazelcast-go-client/internal/hzerror"
+	"github.com/hazelcast/hazelcast-go-client/hzerrors"
 	"github.com/hazelcast/hazelcast-go-client/serialization"
 )
 
@@ -85,7 +85,7 @@ func TestObjectDataInput_AssertAvailable2(t *testing.T) {
 	t.SkipNow()
 	o := NewObjectDataInput([]byte{0, 1, 2, 3}, 3, &Service{}, true)
 	ret := o.AssertAvailable(2)
-	if _, ok := ret.(*hzerror.HazelcastSerializationError); !ok {
+	if _, ok := ret.(*hzerrors.HazelcastSerializationError); !ok {
 		t.Errorf("AssertAvailable() should return error type *hzerror.HazelcastSerializationError but it returns %v", reflect.TypeOf(ret))
 	}
 }
